smv: Add rowwise tiling to the batch norm op.
The reason that drives me to add this change is because for ImageNet based
models, the input size for early layers are pretty big, and without rowwise
tiling, the batch norm op can end up having many tiles with small number of
channels, which makes the data shuffling (e.g., copy data to the tiles) very
expensive. Take the one of the first batch norm layers in ResNet50 as an
example, the input shape is (1, 56, 56, 256), which previously would be tiled
into tiles of shape (1, 56, 32, 8). The small number of channels results in high
overheads on copying data from the original tensor to the tiles, because the
memcpy has very little contiguous data to work with.

This commit adds rowwise tiling, so that the batch norm op supports two more
tiling options: DimNH and DimNHW. Using the same example, now the input tensor
will be tiled into tiles of shape (1, 8, 8, 256), which makes the data shuffling
a lot faster.

TESTED=unit

Change-Id: I3f6e49dbabe5fbbc546d13b27bd70af73d6df97f