python: Add an API for constructing LSTMs.
The user can use this LSTM API in Python to create an LSTM layer. Example
of a two-layer LSTMs:

    inputs = input_data(input_tensor)
    // LSTM layer takes two weights.
    lstm_layer0 = LSTM([w0_tensor, u0_tensor], name="lstm0")
    lstm_layer1 = LSTM([w1_tensor, u1_tensor], name="lstm1")
    // LSTM layer returns outputs and last state.
    outputs, state = lstm_layer0(inputs)
    outputs, state = lstm_layer1(outputs)

The user can also invoke a single step of an LSTM layer. example:

    // Single step invocations of an LSTM layer.
    output_step0, state = lstm_layer0.step(input_step, timestep=0)
    output_step1, state = lstm_layer0.step(input_step, timestep=1)

The implementation unrolls the LSTM layer into timesteps, where each timestep
is further decomposed into the "known" operators, e.g. mat_mul, activation
funcs, elementwise ops, etc. Note that this means the serialized graph proto now
depends on the input time dimension, because the LSTM is effectively unrolled
statically. As a result, the topology file needs to be regenerated even if you
only want to change inputs (aka parameters). A better solution would be adding a
looping operator which dynamically invokes the LSTM step until all the input
timesteps are visited. We will try to add that operator shortly.

This also adds a bidirectional LSTM API which invokes two LSTM layers
underneath.

To make sure the LSTM APIs are doing the correct thing, this also adds a unit
test that validates the SMAUG result directly against the Tensorflow result. It
builds a TF model and creates the same graph topology in SMAUG with parameters
extracted from TF model.

TESTED=unit

Change-Id: I8b792a090fe0739c95df7fb7d7e1a90864ef5146