python,core: Fix proto when a node has multiple output tensors.
The Node proto has parents/children field to indicate the connections between
nodes. To restore the dataflow graph on the C++ side, the network_builder sets
input tensors of a node to the corresponding output tensors of its parent nodes.
It has the assumption that every node only has one output tensor, which isn't
true anymore since we have operators like split. To fix this, we added a
src_tensors_indices field in the Node proto, thus the network_builder can use it
to correctly set the input tensors.

TESTED=unit

Change-Id: I721e310ad0eda6149b759847ffe8f295db29a6c7