python: Change the default naming behavior.
We tried to respect the user's naming choice and raised a KeyError if the
provided name was already used by another node. There seems to have no easy
way for us to maintain this when it comes to adding pseudo-operators. Therefore,
as kicking things down the road, this changes the naming convention to the
simplest, where a suffix will simply be added if a name has been used.

Change-Id: I4ad4449cd837f2c72fcdde080faa2ff47cbcf521