operators: Add a switch operator.
This switch operator is used to forward input tensor to the output
tensor determined by a given pred tensor. For example:

pred = Tensor(data_layout=N,
              tensor_data= np.array([True], dtype=bool))
res_f, res_t = switch(input, pred)

If the value of pred is true, input will be forwarded to res_t. Otherwise,
res_f will be the output.

This commit also introduces the concept of a dead tensor, which is a tensor on
the untaken branch of a control flow. Like in the above example, since pred is
true, res_f will become a dead tensor. When an operator is scheduled, if any of
its inputs is dead, the run() method won't be invoked and instead, its outputs
will be marked as dead tensors.

TESTED=unit

Change-Id: I143bc1251f388d11b6cc0a3e6d4401e2f1fc173f