operator: Fix a bug for compare operators.
We have two factory methods used for creating an operator, defined by
DEF_CREATE_OP(opName, backend) and DEF_CREATE_SMV_OP(opName). The former one
is used for operators that have template specializations for a given backend,
while the latter one allows a custom operator implemeantion for SMV, which is
not limited to explicit template specializations.

In this commit, because these compare operator do have a custom implementation
for SMV, we should use DEF_CREATE_SMV_OP for defining the factory method.

Change-Id: I3b4992a7f0e9412aef5d97025ce4c307da55d2cb