docker: Use release build for LLVM-6.0.
This makes the docker image about 6 GB big, whereas the debug build makes it
35 GB. Except the size, the difference between these two build options is the
release build doesn't provide a clang option to preserve variable names while
the debug build still keeps that. Also, the debug build is noticeably slower.

This makes the docker image use the release build for LLVM-6.0. As the release
build by default passes -discard-value-names to the cc1 compiler and there's
no clang option to avoid that, we need to directly run the raw cc1 command with
the option excluded (this has been done for SMAUG as well as gem5-Aladdin).

Change-Id: I7157435e0a66e8665879c13a5f216fdcb6ef5f73