tests: Move the tests to alongside their tested sources.
We had two types of Python unit tests, one that simply tests attributes of
the protos generated by the Python client APIs, the other actually invokes
the SMAUG C++ core to run the graph proto and validates the result. This
does a bit cosmetics to the tests and moves the second type of tests to their
tested source files.

All tests are still passing.

Change-Id: Iac5b1ac60f0c41b0a936b7a35921379b60e1734e