tests: Move the tests to alongside their tested sources. (#8)
* tests: Move the tests to alongside their tested sources.

We had two types of Python unit tests, one that simply tests attributes of
the protos generated by the Python client APIs, the other actually invokes
the SMAUG C++ core to run the graph proto and validates the result. This
does a bit cosmetics to the tests and moves the second type of tests to their
tested source files.

* python: Purge all import *.

Using import * is easy for the one who writes the code but causes huge
confusion for the user to track definitions. So this gets rid of almost
all import *.

All tests are still passing.