python: Fix a bug in creating unique names.
When creating a unique name for the operator, the naming format is
"%(name)s_%(idx)d". If the name is already in use, we previously just
increased the idx. For example, if "add" is in use, "add_1" will be
used. However, this could mistakenly return a name already in use
as shown in the following example:

  res = add(x, y) # This will be named "add".
  res = add(x, y, name="add_1") # This will be named "add_1".
  res = add(x, y) # BUG: this will also be named "add_1".

To solve this, we keep increasing the idx until the new name is not
in use. In the above example, the third add will now be named "add_2".

This also makes create_unique_queue() public, so the user can get a
unique name before the operator is created. Example:

  res = add(x, y) # This will be named "add".
  # Get a unique name for the operator to be added. Note that
  # `mark_as_used` is set to False in order to avoid re-creating a
  # name for the operator. This returns "add_1".
  unique_name =
      global_vars.get_graph().create_unique_name(
          "add", mark_as_used=False)
  res = add(res, res, name=unique_name) # This will be named "add_1".

TESTED=unit