test: Don't fail the test if protos are not found.
The large ELU network has a very large parameter proto file, which
makes us unable to submit it to github. Instead of exiting the tests
if any input proto is not found, this does a cerr to warn the user
of creating the protos locally and keep on running the rest of the
tests.