core: Allow boost graph to have multiple edges between two vertices.
The graph view of the network didn't allow multiple edges between two vertices.
This leads to inputs unconstructed when an operator's inputs come from the same
input tensor, because the network builder only set an input for an operator if
there's an in_edge. This fixes the issue by making the boost graph use
std::vector to store the edges (instead of std::set), which thus can represent
multiple edges.