python: Add an FP precision test and relax test accuracy. (#19)
The way we model FP16 operations in Aladdin, which converts FP16 to FP32 for
computation and converts the results back to FP16 can potentially lead to
slight precision loss. The aggregated effect can lead to some tests failing
the required accuracy. This adds a test for all the precision-losing operators
against TF operators and also relaxes the required accuracy from 3 decimals to 2
decimals.

This fixes issue #11.

TESTED=unit