python: Simplify reorder op deduplication.
We previously performed a graph-level pass to remove extraneous reorder ops
that were added to transform tensor layouts. This gets rid of the complicated
post-processing of the graph and instead just search the targets of the tensor
to see if there's an existing matched reorder op.

TESTED=all tests are still passing.