docs: Start adding doxygen-compatible doc comments. (#36)
This adds a Doxyfile, a placeholder intro page, and starts the process of making all of our existing documentation Doxygen-compatible. Currently, for C/C++ code, the convention is to prefer adding Doxygen documentation to the header files. Docs can be added to .cpp if it is more appropriate and will not clash with another header file.