Remove git-diff detection logic. (#54)
CircleCI runs all its bash commands with -eo pipefail, so any command
that returns a non-zero exit code will cause the job to fail. In this
case, git-diff will return 1 if there are diffs. The solution is to
ignore git-diff and just create a commit, swallowing the exit code
(which is 1 if there are no diffs). If there is no diff, there will be
no commit, but git push will still return 0.