Add brief descriptions for everything (#66)
This produces a short description of the method/class/variable in the
listing, and only if a more detailed description exists will a larger
block for it be generated further down. This is mostly automatic, but
sometimes Doxygen needs a bit more help.

Also, removes the ugly DOXYGEN_SHOULD_SKIP_THIS macro - this was only
used for Exhale.