tests: Add simulation sweeps in CircleCI build.
We previously added the support in CircirCI for downloading the latest
gem5 binary from gem5-aladdin repo, but didn't actually use the
downloaded binary. This invokes the SMAUG sweep generator and runs the
generated sweeps using the downloaded gem5 binary.

We currently sweep the number of accelerators (1-8) and the SoC interface
choices (DMA or ACP).