operator: createAllTensors() is only used by tests. (#73)
createAllTensors() used to be called by the C++ network builder, but now
all of that logic is handled on the Python side, written into the
network topology proto, and all tensors are created in the network
builder directly. However, unit tests still call this function, so we
leave it as an option, but mark that it's only for testing purposes.