docs: Tutorial for adding new Python operator.
This tutorial builds on the C++ tutorial, which built an operator called
MyCustomOperator. Since the graph and node protos are the link between
Python and C++, this commit also has Doxygen index protobuf files. It
won't parse any of the syntax, but at least it will appear in the list
of files so we can link to the source code.