core: Remove the AllCache memory policy. (#105)
AllCache as a memory policy doesn't make sense. A MemoryPolicy specifies
*how* the data is moved, not where it gets stored after it is moved.
The latter is specified in the Aladdin config files (scratchpads or
caches). This option, when used, causes simulations to fail because
the logic to handle this case was not added for
HybridDatapath::handleCacheMemoryOp; however, in practice it is almost
the same as using ACP, which *does* handle it. So, we just drop this
policy as an option altogether.

Fixes issue #102.